refactor: move result models to namespace ILIAS\Test\Results\Data
